HUB-737: Rename select_documents_variant
This replaces current select_documents controller and associated test
with the variant.
