+++ app/assets/javascripts/application.js.erb
@@ -13,7 +13,6 @@
 
 window.GOVUK.validation.init();
 window.GOVUK.selectDocuments.init();
-window.GOVUK.selectDocumentsVariantC.init();
 window.GOVUK.willItWorkForMe.init();
 window.GOVUK.feedback.init();
 window.GOVUK.furtherInformation.init();
