--- app/views/select_documents_variant_c/index.html.erb
@@ -1,63 +1,0 @@
-<%= page_title 'hub_variant_c.select_documents.title' %>
-<% content_for :feedback_source, 'SELECT_DOCUMENTS_PAGE' %>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-
-    <%= render 'shared/form-errors', errors: flash[:errors], form: @form %>
-
-    <%= form_for @form, url: select_documents_submit_path, html: {id: 'validate-select-documents', novalidate: 'novalidate', 'data-msg': t('hub_variant_c.select_documents.errors.no_selection')} do |f| %>
-      <div class="govuk-form-group<% if @form.errors.include?(:has_valid_passport) %> govuk-form-group--error<% end %>">
-        <fieldset class="govuk-fieldset">
-          <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
-            <h1 class="govuk-fieldset__heading"><%= t 'hub_variant_c.select_documents.heading' %></h1>
-          </legend>
-          <p><%= t 'hub_variant_c.select_documents.explanation' %></p>
-
-          <p><%= t 'hub_variant_c.select_documents.select_all_that_apply' %></p>
-
-          <% if @form.errors.include?(:has_valid_passport) %>
-            <span class="govuk-error-message">
-              <%= @form.errors[:has_valid_passport].first %>
-            </span>
-          <% end %>
-
-          <div id="checkboxes" class="govuk-checkboxes">
-            <div class="govuk-checkboxes__item">
-              <%= f.check_box :has_valid_passport,{class: "govuk-checkboxes__input", id: "select_documents_variant_c_form_has_valid_passport" }, "t", "f" %>
-              <%= f.label :has_valid_passport,  t('hub_variant_c.select_documents.has_valid_passport'), class: "govuk-label govuk-checkboxes__label"%>
-            </div>
-            <div class="govuk-checkboxes__item">
-              <%= f.check_box :has_driving_license,{class: "govuk-checkboxes__input" }, "t", "f" %>
-              <%= f.label :has_driving_license,  t('hub_variant_c.select_documents.has_driving_license'), class: "govuk-label govuk-checkboxes__label"%>
-            </div>
-            <div class="govuk-checkboxes__item">
-              <%= f.check_box :has_phone_can_app,{class: "govuk-checkboxes__input" }, "t", "f" %>
-              <%= f.label :has_phone_can_app,  t('hub_variant_c.select_documents.has_phone_can_app'), class: "govuk-label govuk-checkboxes__label"%>
-            </div>
-            <div class="govuk-checkboxes__item">
-              <%= f.check_box :has_credit_card,{class: "govuk-checkboxes__input" }, "t", "f" %>
-              <%= f.label :has_credit_card,  t('hub_variant_c.select_documents.has_credit_card'), class: "govuk-label govuk-checkboxes__label"%>
-            </div>
-            <div class="govuk-radios__divider">or</div>
-            <div class="govuk-checkboxes__item">
-              <%= f.check_box :has_nothing,{class: "govuk-checkboxes__input" }, "t", "f" %>
-              <%= f.label :has_nothing,  t('hub_variant_c.select_documents.has_nothing'), class: "govuk-label govuk-checkboxes__label"%>
-            </div>
-          </div>
-        </fieldset>
-
-        <details id="progressive_disclosure"  piwik_event_tracking="progressive_disclosure" class="govuk-details govuk-!-padding-top-6" data-module="govuk-details">
-          <summary class="govuk-details__summary">
-            <span class="govuk-details__summary-text"><%= t 'hub_variant_c.select_documents.further_explanation.title' %></span>
-          </summary>
-          <div class="govuk-details__text">
-            <%= t 'hub_variant_c.select_documents.further_explanation.content_html' %>
-          </div>
-        </details>
-
-        <%= f.submit t('navigation.continue'), class: 'govuk-button', id: 'next-button' %>
-      </div>
-    <% end %>
-  </div>
-</div>
