+++ config/main_routes.rb
@@ -136,11 +136,11 @@ get "about", to: "about_loa2#index", as: :about
 get "about_choosing_a_company", to: "about_loa2#choosing_a_company", as: :about_choosing_a_company
 get "will_it_work_for_me", to: "will_it_work_for_me#index", as: :will_it_work_for_me
 post "will_it_work_for_me", to: "will_it_work_for_me#will_it_work_for_me", as: :will_it_work_for_me_submit
-get "select_documents", to: "select_documents_variant_c#index", as: :select_documents
-get "select_documents_none", to: "select_documents_variant_c#no_documents", as: :select_documents_no_documents
-post "select_documents", to: "select_documents_variant_c#select_documents", as: :select_documents_submit
-get "select_documents_advice", to: "select_documents_variant_c#advice", as: :select_documents_advice
-get "prove_your_identity_another_way", to: "select_documents_variant_c#prove_your_identity_another_way", as: :prove_your_identity_another_way
+get "select_documents", to: "select_documents#index", as: :select_documents
+#get "select_documents_none", to: "select_documents#no_documents", as: :select_documents_no_documents
+post "select_documents", to: "select_documents#select_documents", as: :select_documents_submit
+get "select_documents_advice", to: "select_documents#advice", as: :select_documents_advice
+get "prove_your_identity_another_way", to: "select_documents#prove_your_identity_another_way", as: :prove_your_identity_another_way
 get "choose_a_certified_company", to: "choose_a_certified_company_loa2_variant_c#index", as: :choose_a_certified_company
 get "choose_a_certified_company/:company", to: "choose_a_certified_company_loa2_variant_c#about", as: :choose_a_certified_company_about
 post "choose_a_certified_company", to: "choose_a_certified_company_loa2_variant_c#select_idp", as: :choose_a_certified_company_submit
