+++ spec/models/select_documents_form_spec.rb
@@ -1,128 +1,24 @@
-require "spec_helper"
 require "rails_helper"
-
-describe SelectDocumentsForm do
-  context "#validations" do
-    context "#invalid form" do
-      any_driving_license_error_message = "Any driving licence true " + I18n.t("hub.select_documents.errors.no_driving_license_selection")
-      passport_error_message = "Passport true " + I18n.t("hub.select_documents.errors.no_passport_selection")
-      driving_license_issuer_error_message = "Driving licence great britain " + I18n.t("hub.select_documents.errors.no_driving_licence_issuer_selection")
-
-      it "should be invalid if all inputs are empty" do
-        form = SelectDocumentsForm.new({})
-        expect(form).to_not be_valid
-        expect(form.errors.full_messages).to include any_driving_license_error_message
-        expect(form.errors.full_messages).to include passport_error_message
-        expect(form.errors.full_messages).not_to include driving_license_issuer_error_message
-      end
-
-      it "should be invalid if no driving licence issuer details are given" do
-        form = SelectDocumentsForm.new(
-          any_driving_licence: "true",
-          passport: "false",
-        )
-        expect(form).to_not be_valid
-        expect(form.errors.full_messages).to eql [driving_license_issuer_error_message]
-      end
-
-      it "should be invalid if user only inputs driving licence details" do
-        form = SelectDocumentsForm.new(
-          any_driving_licence: "true",
-          driving_licence: "great_britain",
-        )
-        expect(form).to_not be_valid
-        expect(form.errors.full_messages).to eql [passport_error_message]
-      end
-
-      it "should be invalid if user only inputs passport details" do
-        form = SelectDocumentsForm.new(
-          passport: "true",
-        )
-        expect(form).to_not be_valid
-        expect(form.errors.full_messages).to eql [any_driving_license_error_message]
-      end
-    end
-
-    context "#valid form" do
-      it "should be valid if answers are given to every question" do
-        form = SelectDocumentsForm.new(
-          any_driving_licence: "false",
-          passport: "true",
-        )
-        expect(form).to be_valid
-      end
+RSpec.describe SelectDocumentsForm, skip_before: true do
+  context "is invalid" do
+    it "has all check boxes empty" do
+      form = SelectDocumentsForm.new({})
+      expect(form).to_not be_valid
+      expect(form.errors.full_messages.first).to include I18n.t("hub.select_documents.errors.no_selection")
     end
   end
-
-  context "#selected_answers" do
-    it "should return a hash of the selected answers" do
-      form = SelectDocumentsForm.new(
-        passport: "false",
-      )
-      evidence = form.selected_answers
-      expect(evidence).to eql(passport: false)
+  context "is valid" do
+    it "has any check box i.e (has valid passport)" do
+      form = SelectDocumentsForm.new(has_valid_passport: "true")
+      expect(form).to be_valid
     end
-
-    it "should return a hash of the no driving licence and no ni driving licence if no selected for any driving licence" do
-      form = SelectDocumentsForm.new(
-        any_driving_licence: "false",
-      )
-      evidence = form.selected_answers
-      expect(evidence).to eql(driving_licence: false, ni_driving_licence: false)
+    it "has any check box i.e (has nothing) selected" do
+      form = SelectDocumentsForm.new(has_nothing: "true")
+      expect(form).to be_valid
     end
-
-    it "should return a hash of driving licence true if GB driving licence selected" do
-      form = SelectDocumentsForm.new(
-        any_driving_licence: "true",
-        driving_licence: "great_britain",
-      )
-      evidence = form.selected_answers
-      expect(evidence).to eql(driving_licence: true, ni_driving_licence: false)
-    end
-
-    it "should not return selected answers when there it is not an eligible IDP evidence " do
-      form = SelectDocumentsForm.new(
-        passport: "true",
-        any_driving_licence: "",
-      )
-      answers = form.selected_answers
-      expect(answers).to eql(passport: true)
-    end
-  end
-
-  context "#further identity information" do
-    it "should require further information when user has neither uk passport or driving licence" do
-      form = SelectDocumentsForm.new(
-        any_driving_licence: "false",
-        passport: "false",
-      )
-      expect(form).to be_further_id_information_required
-    end
-
-    it "should not require further information when user has a northern ireland driving licence" do
-      form = SelectDocumentsForm.new(
-        any_driving_licence: "true",
-        driving_licence: "northern_ireland",
-        passport: "false",
-      )
-      expect(form).to_not be_further_id_information_required
-    end
-
-    it "should not require further information when user has a GB driving licence" do
-      form = SelectDocumentsForm.new(
-        any_driving_licence: "true",
-        driving_licence: "great_britain",
-        passport: "false",
-      )
-      expect(form).to_not be_further_id_information_required
-    end
-
-    it "should not require further information when user has a UK passport" do
-      form = SelectDocumentsForm.new(
-        any_driving_licence: "false",
-        passport: "true",
-      )
-      expect(form).to_not be_further_id_information_required
+    it "has any check box i.e (has nothing) unselected" do
+      form = SelectDocumentsForm.new(has_nothing: "false")
+      expect(form).to be_valid
     end
   end
 end
